% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{build_size_datasets_plan}
\alias{build_size_datasets_plan}
\title{Make a plan to get size datasets}
\usage{
build_size_datasets_plan(
  datasets_plan,
  datasets_dir = "C:/Users/diaz.renata/Documents/GitHub/BBSsize/analysis/isd_data"
)
}
\arguments{
\item{datasets_plan}{abund plan}

\item{datasets_dir}{where the size datasets live}
}
\value{
a plan
}
\description{
Make a plan to get size datasets
}
