% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{get_mean_e}
\alias{get_mean_e}
\title{Get mean E TS}
\usage{
get_mean_e(abund_ts, energy_ts)
}
\arguments{
\item{abund_ts}{abund}

\item{energy_ts}{energy}
}
\value{
MATSS dataset of mean variable
}
\description{
Get mean E TS
}
