% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{get_size_dataset}
\alias{get_size_dataset}
\title{Load a dataset from BBSSize}
\usage{
get_size_dataset(
  dataset_name,
  datasets_dir = "C:/Users/diaz.renata/Documents/GitHub/BBSsize/analysis/isd_data",
  currency
)
}
\arguments{
\item{dataset_name}{eg "bbs_rtrg_101_4"}

\item{datasets_dir}{where they are}

\item{currency}{"size" or "energy"}
}
\value{
dataset if it exists
}
\description{
Load a dataset from BBSSize
}
