% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{get_species_ts}
\alias{get_species_ts}
\title{Get a species level TS}
\usage{
get_species_ts(abund_table, species_index, currency, route = NA, region = NA)
}
\arguments{
\item{abund_table}{for abundance or energy}

\item{species_index}{which column}

\item{currency}{individuals, energy, or mass}

\item{route}{route}

\item{region}{region}
}
\value{
dataframe of time and population
}
\description{
Get a species level TS
}
