% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{get_sv_ts}
\alias{get_sv_ts}
\title{Get state variable TS}
\usage{
get_sv_ts(a_dataset, currency = "individuals")
}
\arguments{
\item{a_dataset}{a MATSS dataset}

\item{currency}{individuals or energy}
}
\value{
dataframe of time and aggregate variable
}
\description{
Get state variable TS
}
