% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trends.R
\name{linear_trend}
\alias{linear_trend}
\title{Fit linear trend}
\usage{
linear_trend(a_ts)
}
\arguments{
\item{a_ts}{columns for time and response}
}
\value{
data frame of slope, p value, and r2 for ols fit to scaled response
}
\description{
Fit linear trend
}
