% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trends.R
\name{linear_trend_populations}
\alias{linear_trend_populations}
\title{Fit linear trend to all populations in a community}
\usage{
linear_trend_populations(a_dataset)
}
\arguments{
\item{a_dataset}{columns for time and response}

\item{currency}{ind or energy}
}
\value{
data frame of slope, p value, and r2 for ols fit to scaled response for all populations
}
\description{
Fit linear trend to all populations in a community
}
