% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{scale_sv_ts}
\alias{scale_sv_ts}
\title{Scale state variable TS}
\usage{
scale_sv_ts(a_dataset)
}
\arguments{
\item{a_dataset}{a svS dataset}
}
\value{
dataframe of time and aggregate variable
}
\description{
Scale state variable TS
}
