% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{get_all_species_ts}
\alias{get_all_species_ts}
\title{Get species level TSs for a community}
\usage{
get_all_species_ts(a_dataset, currency = "individuals")
}
\arguments{
\item{a_dataset}{a MATSS dataset}

\item{currency}{individuals or energy}
}
\value{
list of dataframes of time and population
}
\description{
Get species level TSs for a community
}
